Handle commit with an empty UTxO
This change was actually quite involved and required other changes in
the API for Server and Games because with Hydra, a commit is a UTxO
and not anymore a single value. This should be somehow abstracted away
and shows the Input interface and Server interface leak abstractions.

Crafting a commit tx was cumbersome as it required interaction with
the HTTP server in Hydra, signing using the cardano key, then
submission which is handled through the cardano-cli even though it
might make sense to use the HTTP server.

The good news is that it seems to work and I can see the output from
the node saying the funds have been committed, which should probably
be followed by collectCom but I cannot find it in the logs so perhaps
this is a bug caused by the fact the node crashed and when it restarts
it does not decide to collect com?