Sketch newGame transaction
Compute elements of the transaction then try to submit it manually
using build-raw to generate the file, then `NewTx` command with the
CBOR body. The transaction gets built and goes through but it fails
beacuse the collateral input does contain non-ADA assets :facepalm:

I was thinking I could just split the utxo in a ada and non-ADA part
but this requires consuming the committed script which requires
collateral.